[Refactor] Unicorn.
Move subscribe handler hpp file out.
Build fix.
