+++ unicorn/include/cocaine/detail/unicorn/zookeeper/subscribe.hpp
@@ -1,0 +1,69 @@
+/*
+* 2015+ Copyright (c) Anton Matveenko <antmat@yandex-team.ru>
+* All rights reserved.
+*
+* This program is free software; you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation; either version 2 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+* GNU General Public License for more details.
+*/
+
+#pragma once
+
+#include "cocaine/unicorn/api.hpp"
+#include "cocaine/unicorn/api/zookeeper.hpp
+
+#include "cocaine/zookeeper/handler.hpp"
+
+namespace cocaine {
+
+/**
+* Action for handling get requests during subscription.
+* When client subscribes for a path - we make a get request to ZK with setting watch on specified path.
+* On each get completion we compare last sent verison to client with current and if current version is greater send update to client.
+* On each watch invoke we issue get command (to later process with this handler) with new watcher.
+*/
+struct subscribe_action_t :
+public zookeeper::managed_data_handler_base_t,
+public zookeeper::managed_watch_handler_base_t,
+public zookeeper::managed_stat_handler_base_t {
+    typedef api::unicorn_t::writable_ptr writable_ptr;
+
+    subscribe_action_t(const zookeeper::handler_tag& tag,
+                       writable_ptr::subscribe _result,
+                       const unicorn::zookeeper_t::context_t& ctx,
+                       unicorn::path_t _path);
+
+    /**
+    * Handling get events
+    */
+    virtual
+    void
+    operator()(int rc, std::string value, const zookeeper::node_stat& stat);
+
+    /**
+    * Handling exist events in case node does not exist.
+    */
+    virtual
+    void
+    operator()(int rc, zookeeper::node_stat const& stat);
+
+    /**
+    * Handling watch events
+    */
+    virtual
+    void
+    operator()(int type, int state, zookeeper::path_t path);
+
+    writable_ptr::subscribe result;
+    unicorn::zookeeper_t::context_t ctx;
+    std::mutex write_lock;
+    unicorn::version_t last_version;
+    const unicorn::path_t path;
+};
+}
